Move and rename K8sNameField component
