+++ core-ui/src/components/K8sNameInput/K8sNameInput.js
